% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_data_doc.R
\docType{data}
\name{crime_data}
\alias{crime_data}
\title{Crime Data of Teenagers}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{\code{burglary}}{integer Counts for burglary crime}
  \item{\code{stolen-property}}{integer Counts for stolen-property crime}
  \item{\code{weapons}}{integer Counts for weapons crime}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
crime_data
}
\description{
This dataset contains offense counts in CA from 1991 to 2010 for burglary, stolen-property and weapons, for offenders between the age of 13 and 14.
}
\keyword{datasets}
