% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_df_doc.R
\docType{data}
\name{earthquake_df}
\alias{earthquake_df}
\title{Earthquake Dataset}
\format{
A data frame with 228 rows and 2 variables:

\describe{
\item{\code{mag}}{double The magnitude of the earthquake}
\item{\code{place}}{character The place where the earthquake occurred}
}
}
\usage{
earthquake_df
}
\description{
This dataset contains a list of earthquakes that occurred across the world during January 2010, and gives their location and magnitude.
}
\keyword{datasets}
