% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileasmatrix.R
\name{fileasmatrix}
\alias{fileasmatrix}
\title{File as matrix}
\usage{
fileasmatrix(infile)
}
\value{
Returns to the matrix of the data
}
\description{
This function takes in a file and convert it into a matrix.The rows with duplicated row names will be dropped.
}
