% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPCA_df.R
\docType{data}
\name{GNPCA_df}
\alias{GNPCA_df}
\title{Annual Real Gross National Product of the U.S.}
\format{
A data frame with 91 rows and 2 variables:
\describe{
  \item{\code{Date}}{character The year of the data measured}
  \item{\code{Value}}{character The value of the annual real gross national product}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GNPCA_df
}
\description{
This dataset contains the Annual Real Gross National Product (GNPCA) of the U.S. from 1929 to 2019.
}
\keyword{datasets}
